% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{absence_one_term_data}
\alias{absence_one_term_data}
\title{Department for Education: Autumn Term Absence Data}
\format{A data frame
\describe{
  \item{term}{Term of which pupil absence relates to, e.g. Autumn, Spring, Summer, Full year}
  \item{year}{Academic year which data is from}
  \item{level}{Describes at which level data is observed, e.g. National level, regional level, Local Authority level, School level}
}}
\source{
\url{https://www.gov.uk/government/statistics/pupil-absence-in-schools-in-england-autumn-term-2017}
}
\usage{
absence_one_term_data
}
\description{
A dataset containing the absence statistics
 for schools in England
}
\keyword{datasets}
