% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_consistency_absence_one_term.R
\name{check_consistency_absence_one_term}
\alias{check_consistency_absence_one_term}
\title{minimal tidy data set for regregrap report production.}
\usage{
check_consistency_absence_one_term(x, log_level = futile.logger::WARN,
  log_appender = "console")
}
\arguments{
\item{x}{Input dataframe, see details.}

\item{log_level}{The severity level at which log messages are written from
least to most serious: TRACE, DEBUG, INFO, WARN, ERROR, FATAL. Default is
level is INFO. See \code{?flog.threshold()} for additional details.}

\item{log_appender}{Defaults to write the log to "console", alternatively you
can provide a character string to specify a filename to also write to. See
for additional details \code{?futile.logger::appender.tee()}.}
}
\value{
If the class is not instantiated correctly, nothing is returned.
}
\description{
\code{check_consistency_absence_one_term} is the class used for the creation of all
 figures and tables in the made up RAP report (it's a demo for the RAP MOOC).
}
\details{
The \code{check_consistency_absence_one_term} class expects a \code{data.frame} with at
least four columns: term, level, year, sess_overall.

  Once inititated, the class has five slots: \code{df}: the basic
  \code{data.frame}, \code{colnames}: a character vector containing the
  column names from the \code{df},
}
\examples{

library(dferap)

df <- check_consistency_absence_one_term(absence_one_term_data)$df

}
